refactor: move test mocks into separate files
